requirements.txt
# cubeSat_sim/requirements.txt

# MQTT клиент для сервисов
paho-mqtt==1.6.1

# Для Pi Camera
picamera==1.13